style: update artifact chart header to have verify button on left
