+++ packages/frontend/src/ee/features/aiCopilot/components/ChatElements/AiChartQuickOptions.tsx
@@ -3,19 +3,10 @@ import {
     type AiArtifact,
     type SavedChart,
 } from '@lightdash/common';
-import {
-    ActionIcon,
-    Button,
-    Group,
-    Menu,
-    Text,
-    Tooltip,
-} from '@mantine-8/core';
+import { ActionIcon, Menu } from '@mantine-8/core';
 import { useDisclosure } from '@mantine-8/hooks';
 import {
     IconChartBar,
-    IconCircleCheck,
-    IconCircleCheckFilled,
     IconDeviceFloppy,
     IconDots,
     IconExternalLink,
@@ -31,12 +22,7 @@ import { useVisualizationContext } from '../../../../../components/LightdashVisu
 import useApp from '../../../../../providers/App/useApp';
 import useTracking from '../../../../../providers/Tracking/useTracking';
 import { EventName } from '../../../../../types/Events';
-import { useSetArtifactVersionVerified } from '../../hooks/useAiAgentArtifacts';
-import { useAiAgentPermission } from '../../hooks/useAiAgentPermission';
-import {
-    useProjectAiAgent,
-    useSavePromptQuery,
-} from '../../hooks/useProjectAiAgents';
+import { useSavePromptQuery } from '../../hooks/useProjectAiAgents';
 import { getOpenInExploreUrl } from '../../utils/getOpenInExploreUrl';
 
 type Props = {
@@ -56,17 +42,12 @@ export const AiChartQuickOptions = ({
     saveChartOptions = { name: '', description: '', linkToMessage: true },
     message,
     compiledSql,
-    artifactData,
 }: Props) => {
     const { track } = useTracking();
     const { user } = useApp();
     const { agentUuid } = useParams();
 
     const [opened, { open, close }] = useDisclosure(false);
-    const [
-        verifyModalOpened,
-        { open: openVerifyModal, close: closeVerifyModal },
-    ] = useDisclosure(false);
     const {
         visualizationConfig,
         columnOrder,
@@ -80,20 +61,8 @@ export const AiChartQuickOptions = ({
         message.threadUuid,
         message.uuid,
     );
-    const { mutate: setVerified } = useSetArtifactVersionVerified(
-        projectUuid,
-        agentUuid!,
-    );
-    const canManageAgent = useAiAgentPermission({
-        action: 'manage',
-        projectUuid,
-    });
-    const { data: agent } = useProjectAiAgent(projectUuid, agentUuid);
     const metricQuery = resultsData?.metricQuery;
     const type = chartConfig.type;
-
-    const isVerified = artifactData?.verifiedByUserUuid !== null;
-    const isAgentVersion3 = agent?.version === 3;
 
     const isDisabled = !metricQuery || !type || !visualizationConfig;
     const onSaveChart = (savedData: SavedChart) => {
@@ -166,59 +135,10 @@ export const AiChartQuickOptions = ({
         }
     };
 
-    const handleVerifyToggle = () => {
-        if (!artifactData) return;
-
-        if (isVerified) {
-            openVerifyModal();
-        } else {
-            setVerified({
-                artifactUuid: artifactData.artifactUuid,
-                versionUuid: artifactData.versionUuid,
-                verified: true,
-            });
-        }
-    };
-
-    const handleConfirmUnverify = () => {
-        if (!artifactData) return;
-        setVerified({
-            artifactUuid: artifactData.artifactUuid,
-            versionUuid: artifactData.versionUuid,
-            verified: false,
-        });
-        closeVerifyModal();
-    };
-
     if (!metricQuery) return null;
 
     return (
         <Fragment>
-            {artifactData && canManageAgent && isAgentVersion3 && (
-                <Tooltip
-                    label={
-                        isVerified
-                            ? 'Remove from verified answers'
-                            : 'Add to verified answers'
-                    }
-                >
-                    <ActionIcon
-                        size="xs"
-                        variant="subtle"
-                        color={isVerified ? 'green' : 'gray'}
-                        onClick={handleVerifyToggle}
-                    >
-                        <MantineIcon
-                            icon={
-                                isVerified
-                                    ? IconCircleCheckFilled
-                                    : IconCircleCheck
-                            }
-                            size="lg"
-                        />
-                    </ActionIcon>
-                </Tooltip>
-            )}
             <Menu withArrow>
                 <Menu.Target>
                     <ActionIcon size="sm" variant="subtle" color="gray">
@@ -304,29 +224,6 @@ export const AiChartQuickOptions = ({
                     }}
                     redirectOnSuccess={false}
                 />
-            </MantineModal>
-            <MantineModal
-                opened={verifyModalOpened}
-                onClose={closeVerifyModal}
-                title="Remove from verified answers"
-                icon={IconCircleCheck}
-                size="sm"
-                actions={
-                    <Group gap="sm">
-                        <Button variant="default" onClick={closeVerifyModal}>
-                            Cancel
-                        </Button>
-                        <Button color="red" onClick={handleConfirmUnverify}>
-                            Confirm
-                        </Button>
-                    </Group>
-                }
-            >
-                <Text>
-                    Are you sure you want to remove this answer from verified
-                    answers? It will no longer be used as an example in future
-                    Agent responses.
-                </Text>
             </MantineModal>
         </Fragment>
     );
