+++ packages/frontend/src/ee/features/aiCopilot/components/ChatElements/AiDashboardVisualization.tsx
@@ -3,23 +3,12 @@ import {
     type AiArtifact,
     type ToolDashboardArgs,
 } from '@lightdash/common';
-import {
-    ActionIcon,
-    Box,
-    Card,
-    Group,
-    Stack,
-    Text,
-    Title,
-} from '@mantine-8/core';
-import { useMediaQuery } from '@mantine-8/hooks';
-import { IconX } from '@tabler/icons-react';
+import { Box, Card, Stack, Text } from '@mantine-8/core';
 import { memo, type FC } from 'react';
-import MantineIcon from '../../../../../components/common/MantineIcon';
 import ErrorBoundary from '../../../../../features/errorBoundary/ErrorBoundary';
 import { clearArtifact } from '../../store/aiArtifactSlice';
 import { useAiAgentStoreDispatch } from '../../store/hooks';
-import { AiDashboardQuickOptions } from './AiDashboardQuickOptions';
+import { AiDashboardVisualizationHeader } from './AiDashboardVisualizationHeader';
 import { AiDashboardVisualizationItem } from './AiDashboardVisualizationItem';
 
 type Props = {
@@ -41,7 +30,6 @@ export const AiDashboardVisualization: FC<Props> = memo(
         showCloseButton = true,
     }) => {
         const dispatch = useAiAgentStoreDispatch();
-        const isMobile = useMediaQuery('(max-width: 768px)');
 
         if (!dashboardConfig?.visualizations) {
             return (
@@ -55,34 +43,14 @@ export const AiDashboardVisualization: FC<Props> = memo(
             <Stack gap={0} h="100%">
                 {/* Dashboard Header with Quick Actions */}
                 <Box pb="md">
-                    <Group gap="md" align="start">
-                        <Stack gap={0} flex={1}>
-                            <Title order={5}>{dashboardConfig.title}</Title>
-                            {dashboardConfig.description && (
-                                <Text c="dimmed" size="xs">
-                                    {dashboardConfig.description}
-                                </Text>
-                            )}
-                        </Stack>
-                        <Group gap="sm" display={isMobile ? 'none' : 'flex'}>
-                            <AiDashboardQuickOptions
-                                artifactData={artifactData}
-                                projectUuid={projectUuid}
-                                agentUuid={agentUuid}
-                                dashboardConfig={dashboardConfig}
-                            />
-                            {showCloseButton && (
-                                <ActionIcon
-                                    size="sm"
-                                    variant="subtle"
-                                    color="gray"
-                                    onClick={() => dispatch(clearArtifact())}
-                                >
-                                    <MantineIcon icon={IconX} color="gray" />
-                                </ActionIcon>
-                            )}
-                        </Group>
-                    </Group>
+                    <AiDashboardVisualizationHeader
+                        projectUuid={projectUuid}
+                        agentUuid={agentUuid}
+                        artifactData={artifactData}
+                        dashboardConfig={dashboardConfig}
+                        showCloseButton={showCloseButton}
+                        onClose={() => dispatch(clearArtifact())}
+                    />
                 </Box>
 
                 {/* Scrollable Dashboard Visualizations */}
